+++ chat-components/src/assets/assets.d.ts
@@ -1,4 +1,0 @@
-declare module "*.svg";
-declare module "*.png";
-declare module "*.gif";
-declare module "*.jpg";