+++ chat-components/src/components/footer/common/defaultProps/customFooterControlProps.ts
@@ -1,8 +1,6 @@
+import { AudioNotificationOffIconBase64, AudioNotificationOnIconBase64, TranscriptDownloadIconBase64, TranscriptEmailIconBase64 } from "../../../../assets/Icons";
+
 import { IFooterControlProps } from "../../interfaces/IFooterControlProps";
-import lcwAudioOff from "../../../../assets/imgs/lcwAudioOff.svg";
-import lcwAudioOn from "../../../../assets/imgs/lcwAudioOn.svg";
-import transcriptDownloadIcon from "../../../../assets/imgs/transcriptDownloadIcon.svg";
-import transcriptEmailIcon from "../../../../assets/imgs/transcriptEmailIcon.svg";
 
 export const customFooterControlProps: IFooterControlProps = {
     id: "oc-lcw-footer",
@@ -18,20 +16,20 @@ export const customFooterControlProps: IFooterControlProps = {
     downloadTranscriptButtonProps: {
         id: "oc-lcw-footer-downloadtranscript-button",
         type: "icon",
-        imageIconProps: { src: transcriptDownloadIcon },
+        imageIconProps: { src: TranscriptDownloadIconBase64 },
         ariaLabel: "Download chat transcript",
     },
     emailTranscriptButtonProps: {
         id: "oc-lcw-footer-emailtranscript-button",
         type: "icon",
-        imageIconProps: { src: transcriptEmailIcon },
+        imageIconProps: { src: TranscriptEmailIconBase64 },
         ariaLabel: "Email Transcript",
     },
     audioNotificationButtonProps: {
         id: "oc-lcw-footer-audionotification-button",
         ariaLabel: "Turn sound off",
         toggleAriaLabel: "Turn sound on",
-        imageIconProps: { src: lcwAudioOn },
-        imageToggleIconProps: { src: lcwAudioOff }
+        imageIconProps: { src: AudioNotificationOnIconBase64 },
+        imageToggleIconProps: { src: AudioNotificationOffIconBase64 }
     }
 };