+++ CHANGE_LOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to this project will be documented in this file.## [Unreleased]
 
-## [0.1.1] - 2022-04-22
+## [Unreleased]
 ### Added 
 - Added the flag for toggling the focus on chatbutton component
 - Updated the version of @microsoft/omnichannel-chat-components to @0.1.0-main.3e712be
+- Added more telemetry and broadcast events for proactive chat pane
+- Added support for PostChat Link Mode
+- Refactored telemetry/broadcast events
+- Added lcw event support